move scenario framework to new namespace
