+++ CommandDotNet.Tests/FeatureTests/Arguments/Operands_DefinedAsArgModel_Defaults.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using CommandDotNet.Attributes;
 using CommandDotNet.Models;
-using CommandDotNet.Tests.BddTests.Framework;
+using CommandDotNet.Tests.ScenarioFramework;
 using CommandDotNet.Tests.FeatureTests.Arguments.Models.ArgsAsArgModels;
 using CommandDotNet.Tests.Utils;
 using Xunit.Abstractions;
