+++ CommandDotNet.Tests/FeatureTests/DefaultCommandMethod.cs
@@ -1,6 +1,6 @@
 using CommandDotNet.Attributes;
 using CommandDotNet.Models;
-using CommandDotNet.Tests.BddTests.Framework;
+using CommandDotNet.Tests.ScenarioFramework;
 using CommandDotNet.Tests.Utils;
 using Xunit.Abstractions;
