+++ CommandDotNet.Tests/FeatureTests/NestedCommand.cs
@@ -1,5 +1,5 @@
 using CommandDotNet.Attributes;
-using CommandDotNet.Tests.BddTests.Framework;
+using CommandDotNet.Tests.ScenarioFramework;
 using CommandDotNet.Tests.Utils;
 using Xunit.Abstractions;
