+++ CommandDotNet.Tests/FeatureTests/UsageAppNameStyles.cs
@@ -1,6 +1,6 @@
 using CommandDotNet.Attributes;
 using CommandDotNet.Models;
-using CommandDotNet.Tests.BddTests.Framework;
+using CommandDotNet.Tests.ScenarioFramework;
 using Xunit.Abstractions;
 
 namespace CommandDotNet.Tests.FeatureTests
