+++ CommandDotNet.Tests/ScenarioFramework/IScenarioContext.cs
@@ -1,4 +1,4 @@
-namespace CommandDotNet.Tests.BddTests.Framework
+namespace CommandDotNet.Tests.ScenarioFramework
 {
     public interface IScenarioContext
     {
