+++ CommandDotNet.Tests/ScenarioFramework/ScenarioContext.cs
@@ -1,4 +1,4 @@
-namespace CommandDotNet.Tests.BddTests.Framework
+namespace CommandDotNet.Tests.ScenarioFramework
 {
     public class ScenarioContext: IScenarioContext
     {
