+++ CommandDotNet.Tests/ScenarioFramework/ScenarioThen.cs
@@ -1,7 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 
-namespace CommandDotNet.Tests.BddTests.Framework
+namespace CommandDotNet.Tests.ScenarioFramework
 {
     public class ScenarioThen
     {
